Split tests in multiple classes
