+++ src/test/java/in/pratanumandal/unique4j/Unique4jTest.java
@@ -17,27 +17,20 @@
 
 package in.pratanumandal.unique4j;
 
-import java.io.File;
-import java.io.IOException;
-import java.nio.charset.Charset;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.atomic.AtomicInteger;
-
-import org.apache.commons.io.FileUtils;
+import in.pratanumandal.unique4j.exception.Unique4jException;
 import org.junit.Assert;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
 
-import in.pratanumandal.unique4j.exception.Unique4jException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.concurrent.atomic.AtomicInteger;
 
 public class Unique4jTest {
 	
-	private static final String TEMP_DIR = System.getProperty("java.io.tmpdir");
-	
-	private static final String BASE_APP_ID = "in.pratanumandal.unique4j-mlsdvo-20191511-#j.6";
-	private static final AtomicInteger APP_ID_COUNT = new AtomicInteger();
+	public static final String BASE_APP_ID = "in.pratanumandal.unique4j-mlsdvo-20191511-#j.6";
+	public static final AtomicInteger APP_ID_COUNT = new AtomicInteger();
 
 	@Test
 	public void testUnique4jBasic() throws Unique4jException {
@@ -447,609 +440,6 @@ public class Unique4jTest {
 		};
 		
 		Unique4j unique2 = new Unique4j(APP_ID, false) {
-			@Override
-			protected String sendMessage() {
-				// send message
-				return message;
-			}
-			
-			@Override
-			protected void receiveMessage(String arg0) {
-				// do nothing
-			}
-		};
-
-		try {
-			// try to obtain lock
-			unique1.acquireLock();
-
-			unique2.acquireLock();
-
-			// wait until message is received
-			if (received.isEmpty()) {
-				synchronized (lock) {
-					try {
-						lock.wait();
-					} catch (InterruptedException e) {
-						e.printStackTrace();
-					}
-				}
-			}
-
-			// assert if message is sent correctly
-			Assert.assertEquals(message, received.get(0));
-		} finally {
-			// try to free the locks before exiting program
-			unique1.releaseLock();
-
-			unique2.releaseLock();
-		}
-	}
-	
-	@Test
-	public void testUnique4jList() throws Unique4jException {
-		final String APP_ID = BASE_APP_ID + "-" + APP_ID_COUNT.getAndIncrement();
-		
-		final Object lock = new Object();
-		
-		final List<String> received = new ArrayList<String>();
-		
-		final List<String> messageList = new ArrayList<String>();
-		messageList.add("ijvnfpp389528$#$@520sdf.213sgv6");
-		messageList.add("ijvnfpp389528$#$@520sdf.213sgv7");
-		messageList.add("ijvnfpp389528$#$@520sdf.213sgv8");
-		messageList.add("ijvnfpp389528$#$@520sdf.213sgv9");
-		
-		Unique4j unique1 = new Unique4jList(APP_ID, false) {
-			@Override
-			protected List<String> sendMessageList() {
-				// send null
-				return null;
-			}
-
-			@Override
-			protected void receiveMessageList(List<String> message) {
-				// to assert on main thread
-				received.addAll(message);
-				
-				// notify that message has been received
-				synchronized (lock) {
-					lock.notify();
-				}
-			}
-		};
-		
-		Unique4j unique2 = new Unique4jList(APP_ID, false) {
-			@Override
-			protected List<String> sendMessageList() {
-				// send message list
-				return messageList;
-			}
-
-			@Override
-			protected void receiveMessageList(List<String> message) {
-				// do nothing
-			}
-		};
-
-		try {
-			// try to obtain lock
-			unique1.acquireLock();
-
-			unique2.acquireLock();
-
-			// wait until message is received
-			if (received.isEmpty()) {
-				synchronized (lock) {
-					try {
-						lock.wait();
-					} catch (InterruptedException e) {
-						e.printStackTrace();
-					}
-				}
-			}
-
-			// assert if message is sent correctly
-			Assert.assertEquals(messageList, received);
-		} finally {
-			// try to free the locks before exiting program
-			unique1.releaseLock();
-
-			unique2.releaseLock();
-		}
-	}
-	
-	@Test
-	public void testUnique4jListNewline() throws Unique4jException {
-		final String APP_ID = BASE_APP_ID + "-" + APP_ID_COUNT.getAndIncrement();
-		
-		final Object lock = new Object();
-		
-		final List<String> received = new ArrayList<String>();
-		
-		final List<String> messageList = new ArrayList<String>();
-		messageList.add("C:\\Users\\Pratanu\nMandal\\Desktop\\rptMrExam.pdf");
-		messageList.add("C:\\Users\\Pratanu Mandal\\Desktop\\nptMrExam.pdf");
-		messageList.add("C:\\Users\\Pratanu Mandal\\Desktop\\'rptMrExam.pdf");
-		messageList.add("C:\\Users\\Pratanu Mandal\\Desktop\\rptMrExam.pdf");
-
-		Unique4j unique1 = new Unique4jList(APP_ID, false) {
-			@Override
-			protected List<String> sendMessageList() {
-				// send null
-				return null;
-			}
-
-			@Override
-			protected void receiveMessageList(List<String> message) {
-				// to assert on main thread
-				received.addAll(message);
-				
-				// notify that message has been received
-				synchronized (lock) {
-					lock.notify();
-				}
-			}
-		};
-		
-		Unique4j unique2 = new Unique4jList(APP_ID, false) {
-			@Override
-			protected List<String> sendMessageList() {
-				// send message list
-				return messageList;
-			}
-
-			@Override
-			protected void receiveMessageList(List<String> message) {
-				// do nothing
-			}
-		};
-
-		try {
-			// try to obtain lock
-			unique1.acquireLock();
-
-			unique2.acquireLock();
-
-			// wait until message is received
-			if (received.isEmpty()) {
-				synchronized (lock) {
-					try {
-						lock.wait();
-					} catch (InterruptedException e) {
-						e.printStackTrace();
-					}
-				}
-			}
-
-			// assert if message is sent correctly
-			Assert.assertEquals(messageList, received);
-		} finally {
-			// try to free the locks before exiting program
-			unique1.releaseLock();
-
-			unique2.releaseLock();
-		}
-	}
-	
-	@Test
-	public void testUnique4jListNull() throws Unique4jException {
-		final String APP_ID = BASE_APP_ID + "-" + APP_ID_COUNT.getAndIncrement();
-		
-		final Object lock = new Object();
-		
-		final List<Object> received = new ArrayList<Object>();
-		
-		Unique4j unique1 = new Unique4jList(APP_ID, false) {
-			@Override
-			protected List<String> sendMessageList() {
-				// send null
-				return null;
-			}
-
-			@Override
-			protected void receiveMessageList(List<String> message) {
-				// to assert on main thread
-				received.add(message);
-				
-				// notify that message has been received
-				synchronized (lock) {
-					lock.notify();
-				}
-			}
-		};
-		
-		Unique4j unique2 = new Unique4jList(APP_ID, false) {
-			@Override
-			protected List<String> sendMessageList() {
-				// send null
-				return null;
-			}
-
-			@Override
-			protected void receiveMessageList(List<String> message) {
-				// do nothing
-			}
-		};
-
-		try {
-			// try to obtain lock
-			unique1.acquireLock();
-
-			unique2.acquireLock();
-
-			// wait until message is received
-			if (received.isEmpty()) {
-				synchronized (lock) {
-					try {
-						lock.wait();
-					} catch (InterruptedException e) {
-						e.printStackTrace();
-					}
-				}
-			}
-
-			// assert if message is sent correctly
-			Assert.assertNull(received.get(0));
-		} finally {
-			// try to free the locks before exiting program
-			unique1.releaseLock();
-
-			unique2.releaseLock();
-		}
-	}
-	
-	@Test
-	public void testUnique4jMap() throws Unique4jException {
-		final String APP_ID = BASE_APP_ID + "-" + APP_ID_COUNT.getAndIncrement();
-		
-		final Object lock = new Object();
-		
-		final Map<String, String> received = new HashMap<String, String>();
-		
-		final Map<String, String> messageMap = new HashMap<String, String>();
-		messageMap.put("23vvf1", "ijvnfpp389528$#$@520sdf.213sgv6");
-		messageMap.put("23vvf2", "ijvnfpp389528$#$@520sdf.213sgv7");
-		messageMap.put("23vvf3", "ijvnfpp389528$#$@520sdf.213sgv8");
-		messageMap.put("23vvf4", "ijvnfpp389528$#$@520sdf.213sgv9");
-		
-		Unique4j unique1 = new Unique4jMap(APP_ID, false) {
-			@Override
-			protected Map<String, String> sendMessageMap() {
-				// send null
-				return null;
-			}
-
-			@Override
-			protected void receiveMessageMap(Map<String, String> message) {
-				// to assert on main thread
-				received.putAll(message);
-				
-				// notify that message has been received
-				synchronized (lock) {
-					lock.notify();
-				}
-			}
-		};
-		
-		Unique4j unique2 = new Unique4jMap(APP_ID, false) {
-			@Override
-			protected Map<String, String> sendMessageMap() {
-				// send message list
-				return messageMap;
-			}
-
-			@Override
-			protected void receiveMessageMap(Map<String, String> message) {
-				// do nothing
-			}
-		};
-
-		// try to obtain lock
-		unique1.acquireLock();
-
-		unique2.acquireLock();
-
-		try {
-			// wait until message is received
-			if (received.isEmpty()) {
-				synchronized (lock) {
-					try {
-						lock.wait();
-					} catch (InterruptedException e) {
-						e.printStackTrace();
-					}
-				}
-			}
-
-			// assert if message is sent correctly
-			Assert.assertEquals(messageMap, received);
-		} finally {
-			// try to free the locks before exiting program
-			unique1.releaseLock();
-
-			unique2.releaseLock();
-		}
-	}
-	
-	@Test
-	public void testUnique4jMapNewline() throws Unique4jException {
-		final String APP_ID = BASE_APP_ID + "-" + APP_ID_COUNT.getAndIncrement();
-		
-		final Object lock = new Object();
-		
-		final Map<String, String> received = new HashMap<String, String>();
-		
-		final Map<String, String> messageMap = new HashMap<String, String>();
-		messageMap.put("23vvf1", "C:\\Users\\Pratanu\nMandal\\Desktop\\rptMrExam.pdf");
-		messageMap.put("23vvf2", "C:\\Users\\Pratanu Mandal\\Desktop\\nptMrExam.pdf");
-		messageMap.put("23vvf3", "C:\\Users\\Pratanu Mandal\\Desktop\\'rptMrExam.pdf");
-		messageMap.put("23vvf4", "C:\\Users\\Pratanu Mandal\\Desktop\\rptMrExam.pdf");
-		
-		Unique4j unique1 = new Unique4jMap(APP_ID, false) {
-			@Override
-			protected Map<String, String> sendMessageMap() {
-				// send null
-				return null;
-			}
-
-			@Override
-			protected void receiveMessageMap(Map<String, String> message) {
-				// to assert on main thread
-				received.putAll(message);
-				
-				// notify that message has been received
-				synchronized (lock) {
-					lock.notify();
-				}
-			}
-		};
-		
-		Unique4j unique2 = new Unique4jMap(APP_ID, false) {
-			@Override
-			protected Map<String, String> sendMessageMap() {
-				// send message list
-				return messageMap;
-			}
-
-			@Override
-			protected void receiveMessageMap(Map<String, String> message) {
-				// do nothing
-			}
-		};
-
-		// try to obtain lock
-		unique1.acquireLock();
-
-		unique2.acquireLock();
-
-		try {
-			// wait until message is received
-			if (received.isEmpty()) {
-				synchronized (lock) {
-					try {
-						lock.wait();
-					} catch (InterruptedException e) {
-						e.printStackTrace();
-					}
-				}
-			}
-
-			// assert if message is sent correctly
-			Assert.assertEquals(messageMap, received);
-		} finally {
-			// try to free the locks before exiting program
-			unique1.releaseLock();
-
-			unique2.releaseLock();
-		}
-	}
-	
-	@Test
-	public void testUnique4jMapNull() throws Unique4jException {
-		final String APP_ID = BASE_APP_ID + "-" + APP_ID_COUNT.getAndIncrement();
-		
-		final Object lock = new Object();
-		
-		final List<Object> received = new ArrayList<Object>();
-		
-		Unique4j unique1 = new Unique4jMap(APP_ID, false) {
-			@Override
-			protected Map<String, String> sendMessageMap() {
-				// send null
-				return null;
-			}
-
-			@Override
-			protected void receiveMessageMap(Map<String, String> message) {
-				// to assert on main thread
-				received.add(message);
-				
-				// notify that message has been received
-				synchronized (lock) {
-					lock.notify();
-				}
-			}
-		};
-
-		Unique4j unique2 = new Unique4jMap(APP_ID, false) {
-			@Override
-			protected Map<String, String> sendMessageMap() {
-				// send null
-				return null;
-			}
-
-			@Override
-			protected void receiveMessageMap(Map<String, String> message) {
-				// do nothing
-			}
-		};
-
-		try {
-			// try to obtain lock
-			unique1.acquireLock();
-
-			unique2.acquireLock();
-
-			// wait until message is received
-			if (received.isEmpty()) {
-				synchronized (lock) {
-					try {
-						lock.wait();
-					} catch (InterruptedException e) {
-						e.printStackTrace();
-					}
-				}
-			}
-
-			// assert if message is sent correctly
-			Assert.assertNull(received.get(0));
-		} finally {
-			// try to free the locks before exiting program
-			unique1.releaseLock();
-
-			unique2.releaseLock();
-		}
-	}
-	
-	@Test
-	public void testCorruptedLockFile() throws Unique4jException, IOException {
-		final String APP_ID = BASE_APP_ID + "-" + APP_ID_COUNT.getAndIncrement();
-		
-		String filePath = TEMP_DIR + File.separator + APP_ID + ".lock";
-		
-		File file = new File(filePath);
-		
-		// if lock file exists, delete it
-		if (file.exists()) {
-			FileUtils.forceDelete(file);
-		}
-
-		// create a corrupted lock file
-		FileUtils.writeStringToFile(file, "abcdefghi\njklmnop\n\rqrst", Charset.forName("UTF-8"));
-
-		// create instance of Unique
-		Unique4j unique = new Unique4j(APP_ID) {
-			@Override
-			public String sendMessage() {
-				// send null
-				return null;
-			}
-			
-			@Override
-			public void receiveMessage(String message) {
-				// do nothing
-			}
-		};
-
-		try {
-			// try to obtain lock
-			unique.acquireLock();
-		} finally {
-			// try to free the lock before exiting program
-			unique.releaseLock();
-		}
-	}
-	
-	@Test
-	public void testCustomPortStatic() throws Unique4jException {
-		final String APP_ID = BASE_APP_ID + "-" + APP_ID_COUNT.getAndIncrement();
-		
-		final Object lock = new Object();
-		
-		final List<String> received = new ArrayList<String>();
-		
-		final String message = "ijvnfpp389528$#$@520sdf.213sgv8";
-		
-		Unique4j unique1 = new Unique4j(APP_ID, false, 8080, PortPolicy.STATIC) {
-			@Override
-			protected String sendMessage() {
-				// send null
-				return null;
-			}
-			
-			@Override
-			protected void receiveMessage(String arg0) {
-				// to assert on main thread
-				received.add(arg0);
-				
-				// notify that message has been received
-				synchronized (lock) {
-					lock.notify();
-				}
-			}
-		};
-		
-		Unique4j unique2 = new Unique4j(APP_ID, false, 8080, PortPolicy.STATIC) {
-			@Override
-			protected String sendMessage() {
-				// send message
-				return message;
-			}
-			
-			@Override
-			protected void receiveMessage(String arg0) {
-				// do nothing
-			}
-		};
-
-		try {
-			// try to obtain lock
-			unique1.acquireLock();
-
-			unique2.acquireLock();
-
-			// wait until message is received
-			if (received.isEmpty()) {
-				synchronized (lock) {
-					try {
-						lock.wait();
-					} catch (InterruptedException e) {
-						e.printStackTrace();
-					}
-				}
-			}
-
-			// assert if message is sent correctly
-			Assert.assertEquals(message, received.get(0));
-		} finally {
-			// try to free the locks before exiting program
-			unique1.releaseLock();
-
-			unique2.releaseLock();
-		}
-	}
-	
-	@Test
-	public void testCustomPortDynamic() throws Unique4jException {
-		final String APP_ID = BASE_APP_ID + "-" + APP_ID_COUNT.getAndIncrement();
-		
-		final Object lock = new Object();
-		
-		final List<String> received = new ArrayList<String>();
-		
-		final String message = "ijvnfpp389528$#$@520sdf.213sgv8";
-		
-		Unique4j unique1 = new Unique4j(APP_ID, false, 8080, PortPolicy.DYNAMIC) {
-			@Override
-			protected String sendMessage() {
-				// send null
-				return null;
-			}
-			
-			@Override
-			protected void receiveMessage(String arg0) {
-				// to assert on main thread
-				received.add(arg0);
-				
-				// notify that message has been received
-				synchronized (lock) {
-					lock.notify();
-				}
-			}
-		};
-		
-		Unique4j unique2 = new Unique4j(APP_ID, false, 8080, PortPolicy.DYNAMIC) {
 			@Override
 			protected String sendMessage() {
 				// send message
